% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{close_database}
\alias{close_database}
\title{Close connection to database}
\usage{
close_database(con)
}
\arguments{
\item{con}{A DBI connection object from open_database()}
}
\description{
Close connection to SQLite database. Recommended at the end of every session.
}
