% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{collect_data}
\alias{collect_data}
\title{Collect data into R}
\usage{
collect_data(data)
}
\arguments{
\item{data}{A DBI table that can be treated like a data.frame. See \code{open_fish()} and \code{open_survey()}}
}
\description{
Collect data into R and convert dates/datetimes into the correct data types with the correct time zone.
It is recommended to use this function instead of \code{collect} because the database RSQLite does not
support date and time data types, so they are stored as character vectors.
Although some date and time operations are still posssible, when you \code{collect} the dataset, the
Date and Datetime columns will be character vectors. This function will convert those columns
(if they exist in your collected dataset) into the correct date and datetime format.
}
