% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_fish_db.R
\name{connect_fish_db}
\alias{connect_fish_db}
\title{Connect fish database}
\usage{
connect_fish_db()
}
\value{
conn A database connection object
}
\description{
Connect to the fish database stored in the package
}
\examples{

con <- connect_fish_db()

}
