% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lengths.R
\name{convert_lengths}
\alias{convert_lengths}
\title{Convert fish length}
\usage{
convert_lengths(data)
}
\arguments{
\item{data}{An arrow dataset of fish data. See \code{open_fish()}}
}
\value{
data_conv An arrow dataset with converted lengths
}
\description{
Converts fish length data using the length conversion table. Returns an arrow dataset.
}
\examples{
library(dplyr)

fish <- open_fish() \%>\% 
    filter(Taxa \%in\% c("Dorosoma petenense", "Morone saxatilis", "Spirinchus thaleichthys"))

fish_conv <- convert_lengths(fish) \%>\% 
    collect()

}
