% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_fish_db}
\alias{create_fish_db}
\title{Create fish database}
\usage{
create_fish_db(edi_pid = NULL, update = FALSE)
}
\arguments{
\item{edi_pid}{(char) Optionally, a way to specify a specific revision of the dataset, in the format "edi.1075.1"
Leave parameter unset to get the latest revision.}

\item{update}{(logical) If set to TRUE, will update to latest version from EDI if a newer version is available}
}
\description{
Function to create the fish database. Reads in raw data from the
published \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1075&revision=1}{EDI dataset}.
}
