% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_cache_updated}
\alias{is_cache_updated}
\title{Is cached data up to date with latest EDI data}
\usage{
is_cache_updated(cache_dir = "deltafish")
}
\arguments{
\item{cache_dir}{(char) The cache directory, by default set to deltafish for most use cases.}
}
\value{
(logical) Whether cache is up to date
}
\description{
Returns TRUE for up to date, FALSE if a newer version exists
}
