% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_fish}
\alias{open_fish}
\title{Open fish data}
\usage{
open_fish()
}
\value{
fish An arrow object that can be treated like a data.frame, with fish data
}
\description{
Open the survey data stored in the package
}
