% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_fish_db.R
\name{open_length_conv}
\alias{open_length_conv}
\title{Connect length conversion data}
\usage{
open_length_conv()
}
\value{
length_conv A database connection object
}
\description{
Connect to the length conversion data stored in the package
}
