% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_survey}
\alias{open_survey}
\title{Connect survey data}
\usage{
open_survey()
}
\value{
An arrow object that can be treated like a data.frame, with survey data
}
\description{
Connect to the survey data stored in the package
}
