% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_unknown_lengths.R
\name{remove_unknown_lengths}
\alias{remove_unknown_lengths}
\title{Remove unknown fish lengths}
\usage{
remove_unknown_lengths(data, univariate)
}
\arguments{
\item{data}{A DBI table that can be treated like a data.frame, with fish data. See \code{open_fish()}}

\item{univariate}{(logical) Will these data be used for univariate analyses
(\code{univariate=TRUE})? Or multi-species analyses (\code{univariate=FALSE})?
If univariate, when a \code{Length_NA_flag=="Unknown length"} record is found,
all records of that taxa from that sample are removed. In effect, this is
transforming those records into missing data. If \code{univariate=FALSE}, when
a \code{Length_NA_flag=="Unknown length"} record is found, the entire sample
is removed and no 0s are filled in, since accurate community data cannot be
confirmed for that sample.}
}
\value{
data_known A DBI table with only known lengths
}
\description{
Removes unknown fish lengths. Returns a DBI table.
}
