% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_cache}
\alias{show_cache}
\title{Show list of cached deltafish files}
\usage{
show_cache()
}
\value{
(list) A list of files
}
\description{
This function returns a list of files cached for the package.
}
