% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{show_cached_revision}
\alias{show_cached_revision}
\title{Show revision number of cached files}
\usage{
show_cached_revision(cache_dir = "deltafish")
}
\arguments{
\item{(char)}{The cache directory, by default set to deltafish for most use cases.}
}
\value{
(char) The revision number in the cache
}
\description{
This function returns the EDI revision number of the cached data.
}
