% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_lengths.R
\name{convert_lengths}
\alias{convert_lengths}
\title{Convert fish length}
\usage{
convert_lengths(data)
}
\arguments{
\item{data}{A DBI table that can be treated like a data.frame, with fish data. See \code{open_fish()}}
}
\value{
data_conv A DBI table with converted lengths
}
\description{
Converts fish length data using the length conversion table. Returns an arrow dataset.
This function is only needed to convert Suisun survey data.
}
\examples{
\dontrun{
library(dplyr)
con <- open_database()
fish <- open_fish(con) \%>\%
  filter(Taxa \%in\% c("Dorosoma petenense", "Morone saxatilis", "Spirinchus thaleichthys"))

fish_conv <- convert_lengths(fish) \%>\%
  collect()
close_database(con)
}

}
