% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_fish_db}
\alias{create_fish_db}
\title{Create fish database}
\usage{
create_fish_db()
}
\description{
Function to create the arrow dataset. Reads in raw data from the
published \href{https://portal.edirepository.org/nis/mapbrowse?scope=edi&identifier=1075&revision=1}{EDI dataset}.
}
