% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create_fish_db}
\alias{create_fish_db}
\title{Create fish database}
\usage{
create_fish_db()
}
\description{
Function to create the arrow dataset. Currently hard coded
to look at data stored on the KNB. Eventually will look for data
on EDI. Only used to generate the datasets in a local cache.
}
