% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fish_db.R
\name{create_fish_db}
\alias{create_fish_db}
\title{Create fish database}
\usage{
create_fish_db()
}
\description{
Function to create fish database. Currently hard coded
to look at raw data stored locally. Eventually will look for data
on EDI.
}
