% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deltafish-package.R
\docType{package}
\name{deltafish-package}
\alias{deltafish}
\alias{deltafish-package}
\title{deltafish: Accesses an Integrated Fish Count and Length Dataset from the San Francisco Delta}
\description{
This package enables streamlined access to a large (45 million observation) integrated fish dataset from the San Fransisco Delta. The package downloads published data and stores it in parquet files as an `arrow` dataset in a local cache. Helper functions enable efficient querying of this large dataset.
}
\author{
\strong{Maintainer}: Jeanette Clark \email{jclark@nceas.ucsb.edu} (\href{https://orcid.org/0000-0003-4703-1974}{ORCID})

Authors:
\itemize{
  \item Samuel M Bashevkin \email{sam.bashevkin@deltacouncil.ca.gov} (\href{https://orcid.org/0000-0001-7406-7089}{ORCID})
}

}
\keyword{internal}
