% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_fish}
\alias{open_fish}
\title{Connect fish data}
\usage{
open_fish(con, quiet = FALSE)
}
\arguments{
\item{con}{A DBI connection object from open_database()}

\item{quiet}{silence message about fish length units.}
}
\value{
A DBI table that can be treated like a data.frame, with fish data
}
\description{
Connect to the fish table stored in the database
}
