% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_length_conv}
\alias{open_length_conv}
\title{Connect length conversion data}
\usage{
open_length_conv()
}
\value{
fish An arrow object that can be treated like a data.frame, with length conversion data
}
\description{
Connect to the length conversion data stored in the package
}
