% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.R
\name{open_survey}
\alias{open_survey}
\title{Connect survey data}
\usage{
open_survey(con)
}
\arguments{
\item{con}{A DBI connection object from open_database()}
}
\value{
A DBI table that can be treated like a data.frame, with survey data
}
\description{
Connect to the survey table stored in the database
}
